[REFACTOR] ProductList에서 data fetch 분리 - 컴포넌트를 재활용할 수 있도록 해줍니다.
